backported: possible null pointer dereference fix, and cast
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@13843 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/utils/GUIInfoManager.cpp
@@ -2113,7 +2113,7 @@ bool CGUIInfoManager::GetMultiInfoBool(const GUIInfo &info, DWORD dwContextWindo
       {
         const CGUIViewState *viewState = ((CGUIMediaWindow*)window)->GetViewState();
         if (viewState)
-          bReturn = (viewState->GetSortMethod() == info.GetData1());
+          bReturn = ((unsigned int)viewState->GetSortMethod() == info.GetData1());
       }
       break;
     }
@@ -2124,7 +2124,7 @@ bool CGUIInfoManager::GetMultiInfoBool(const GUIInfo &info, DWORD dwContextWindo
       {
         const CGUIViewState *viewState = ((CGUIMediaWindow*)window)->GetViewState();
         if (viewState)
-          bReturn = (viewState->GetDisplaySortOrder() == info.GetData1());
+          bReturn = ((unsigned int)viewState->GetDisplaySortOrder() == info.GetData1());
       }
       break;
     }
